ensure config data is int

+++ custom_components/solakon_one/modbus.py
@@ -32,7 +32,7 @@ class SolakonModbusHub:
         host: str,
         port: int,
         device_id: int,
-        scan_interval: int,
+        scan_interval: float,
     ) -> None:
         """Initialize the Modbus hub."""
         self._hass = hass
@@ -353,6 +353,6 @@ def get_modbus_hub(hass: HomeAssistant, data: ConfigEntry) -> SolakonModbusHub:
         hass,
         data[CONF_HOST],
         data[CONF_PORT],
-        int(data.get(CONF_DEVICE_ID, DEFAULT_DEVICE_ID)),
+        data.get(CONF_DEVICE_ID, DEFAULT_DEVICE_ID),
         data.get(CONF_SCAN_INTERVAL, DEFAULT_SCAN_INTERVAL),
     )
